requirements.txt
# # Django Web Framework
# Django>=4.2


# # PyAnnote for speaker diarization
# pyannote-audio==3.1.1
# torch>=2.0.0
# torchaudio

# # Whisper API Client (Groq-compatible, use openai or httpx based on integration)
# openai>=1.10.0
# httpx>=0.27.0

# # For sentiment analysis (e.g., using transformers or TextBlob)
# transformers>=4.40.0
# torch>=2.0.0  # already mentioned, required for transformers

# # Optional NLP tools
# nltk
# textblob

# # Visualization (for speaker stats, etc.)
# matplotlib
# seaborn
# plotly

# # General utilities
# numpy
# pandas
# scikit-learn

# # Optional for summarization (e.g., Hugging Face models)
# sentence-transformers
# sumy


# Django Web Framework
Django>=4.2

# Audio Handling & Processing
pydub
imageio-ffmpeg  # needed for setting ffmpeg path
ffmpeg-python   # optional but helpful if you want to control ffmpeg in Python

# PyAnnote for Speaker Diarization
pyannote-audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# Whisper-based API (Groq-compatible)
openai>=1.10.0
httpx>=0.27.0

# # Sentiment Analysis
# nltk
# textblob

# # Optional: More advanced NLP tools (can be used for better sentiment/summary)
# transformers>=4.40.0
# sentence-transformers
# sumy

# # Visualization (optional but good for speaker stats)
# matplotlib
# seaborn
# plotly

# General-purpose utilities
numpy
pandas
scikit-learn
